600 column switcher disappears

+++ classes/ListScreenRepository/Storage/ListScreenRepository.php
@@ -1,38 +1,26 @@
-<?php
+<?php declare( strict_types=1 );
 
 namespace AC\ListScreenRepository\Storage;
 
 use AC;
 use AC\Exception;
 use AC\ListScreen;
+use AC\ListScreenCollection;
 use AC\ListScreenRepository\Rules;
 use AC\ListScreenRepository\SourceAware;
 use AC\Type\ListScreenId;
 use LogicException;
+use WP_User;
 
 class ListScreenRepository implements AC\ListScreenRepositoryWritable, SourceAware {
 
-	/**
-	 * @var AC\ListScreenRepository
-	 */
 	private $repository;
 
-	/**
-	 * @var bool
-	 */
 	private $writable;
 
-	/**
-	 * @var Rules
-	 */
 	private $rules;
 
-	/**
-	 * @param AC\ListScreenRepository $repository
-	 * @param bool|null               $writable
-	 * @param Rules|null              $rules
-	 */
-	public function __construct( AC\ListScreenRepository $repository, $writable = null, Rules $rules = null ) {
+	public function __construct( AC\ListScreenRepository $repository, bool $writable = null, Rules $rules = null ) {
 		if ( null === $writable ) {
 			$writable = false;
 		}
@@ -42,19 +30,11 @@ class ListScreenRepository implements AC\ListScreenRepositoryWritable, SourceAwa
 		$this->rules = $rules;
 	}
 
-	/**
-	 * @return bool
-	 */
-	public function is_writable() {
+	public function is_writable(): bool {
 		return $this->writable;
 	}
 
-	/**
-	 * @param bool $writable
-	 *
-	 * @return self
-	 */
-	public function with_writable( $writable ) {
+	public function with_writable( bool $writable ): self {
 		return new self(
 			$this->repository,
 			$writable,
@@ -62,10 +42,7 @@ class ListScreenRepository implements AC\ListScreenRepositoryWritable, SourceAwa
 		);
 	}
 
-	/**
-	 * @return Rules
-	 */
-	public function get_rules() {
+	public function get_rules(): Rules {
 		if ( ! $this->has_rules() ) {
 			throw new LogicException( 'No rules defined.' );
 		}
@@ -73,14 +50,21 @@ class ListScreenRepository implements AC\ListScreenRepositoryWritable, SourceAwa
 		return $this->rules;
 	}
 
-	/**
-	 * @return bool
-	 */
-	public function has_rules() {
+	public function has_rules(): bool {
 		return $this->rules !== null;
 	}
 
-	public function find( ListScreenId $id ) {
+	public function find_by_user( ListScreenId $id, WP_User $user ): ?ListScreen {
+		$list_screen = $this->repository->find_by_user( $id, $user );
+
+		if ( $list_screen && ! $this->is_writable() ) {
+			$list_screen->set_read_only( true );
+		}
+
+		return $list_screen;
+	}
+
+	public function find( ListScreenId $id ): ?ListScreen {
 		$list_screen = $this->repository->find( $id );
 
 		if ( $list_screen && ! $this->is_writable() ) {
@@ -90,31 +74,59 @@ class ListScreenRepository implements AC\ListScreenRepositoryWritable, SourceAwa
 		return $list_screen;
 	}
 
-	public function exists( ListScreenId $id ) {
-		return $this->repository->exists( $id );
-	}
-
-	public function find_all( array $args = [] ) {
-		$list_screens = $this->repository->find_all( $args );
+	public function find_all_by_key( string $key, string $order_by = null ): ListScreenCollection {
+		$list_screens = $this->repository->find_all_by_key( $key, $order_by );
 
 		if ( ! $this->is_writable() ) {
-			foreach ( $list_screens as $list_screen ) {
-				$list_screen->set_read_only( true );
-			}
+			$this->set_all_read_only( $list_screens );
 		}
 
 		return $list_screens;
 	}
 
-	public function save( ListScreen $list_screen ) {
-		$this->repository->save( $list_screen );
+	public function find_all_by_user( string $key, WP_User $user, string $order_by = null ): ListScreenCollection {
+		$list_screens = $this->repository->find_all_by_user( $key, $user, $order_by );
+
+		if ( ! $this->is_writable() ) {
+			$this->set_all_read_only( $list_screens );
+		}
+
+		return $list_screens;
 	}
 
-	public function delete( ListScreen $list_screen ) {
-		$this->repository->delete( $list_screen );
+	private function set_all_read_only( ListScreenCollection $list_screens ): void {
+		foreach ( $list_screens as $list_screen ) {
+			$list_screen->set_read_only( true );
+		}
 	}
 
-	public function get_source( ListScreenId $id ) {
+	public function exists( ListScreenId $id ): bool {
+		return $this->repository->exists( $id );
+	}
+
+	public function find_all( string $order_by = null ): ListScreenCollection {
+		$list_screens = $this->repository->find_all( $order_by );
+
+		if ( ! $this->is_writable() ) {
+			$this->set_all_read_only( $list_screens );
+		}
+
+		return $list_screens;
+	}
+
+	public function save( ListScreen $list_screen ): void {
+		if ( $this->repository instanceof AC\ListScreenRepositoryWritable ) {
+			$this->repository->save( $list_screen );
+		}
+	}
+
+	public function delete( ListScreen $list_screen ): void {
+		if ( $this->repository instanceof AC\ListScreenRepositoryWritable ) {
+			$this->repository->delete( $list_screen );
+		}
+	}
+
+	public function get_source( ListScreenId $id ): string {
 		if ( ! $this->has_source( $id ) ) {
 			throw new Exception\SourceNotAvailableException();
 		}
@@ -122,7 +134,7 @@ class ListScreenRepository implements AC\ListScreenRepositoryWritable, SourceAwa
 		return $this->repository->get_source( $id );
 	}
 
-	public function has_source( ListScreenId $id ) {
+	public function has_source( ListScreenId $id ): bool {
 		return $this->repository instanceof SourceAware && $this->repository->has_source( $id );
 	}
